Ticket: #1910 Ticket: #1911
Incremental updates for archiving.  Added stored procedures for querying the restorequeue and archivequeue, added some preliminary code for using those procedures when archiving and restoring.  Also, cleaned up a few minor things with the recent checkin to remove the offis toolkit.  Finally, introduced the SharpZipLib library for use of the "tar" methods in this library.