Bug: #3256
Made improvements to the compression codecs.  As part of the testing, also made changes to the dicom tool kit sample application so we could store /export the raw pixel data (whether it be compressed or uncompressed) into a file, to make testing with external tools easier.

After this change, lossy compression is fixed for this image, however, a problem was identified with lossless compression of the same object where decompression was not working properly.  Used the above changes to extract the pixel data after compression, and we could view the j2k data w/ IrfanViewer, but could not decompress it with a stand along OpenJpeg tool.  The resultant image looked the same as when doing decompression from within the dicom library.