Ticket(s): #3654 - Cannot compress Palette Color Images
Schema/Configuration Changes: N/A

Updates to convert (if specified) Palette Color photometric interpretation images to RGB before compression.  Added a new codec parameter, which also can be set in the rules to specify if the conversion should be conducted.  Also made a fix to the PaletteColorLut class with checking parameters.