[Defect #10046] - ShredHost service repeatedly crashes when auto-routing certain already compressed images
Did a code review of the jpeg compression fix.  Found a few places where the error
handling with the new try/catch blocks could be better.  Also found a place
related to the "preferred transfer syntax" code done for the IV that codec
exceptions were not be caught properly.  Finally, cleaned up some formatting in
a few places and added support for a few more sop classes to the sample app
for testing purposes.  Also, recompiled the jpeg codecs (x86 & x64) after
the changes and checked them in.